Added features to support level bot in common libs.
Added support to pickup items worth more than 10 gold/space and pickup all gold if total gold is < Config.LowGold.
Changed identify and clearInventory functions to sell this junk.
Changed clearInventory to not sell/drop scrolls if the corresponding tome is not owned.

(Was originally based on master branch, moved it manually over to TestVer)