Minor changes to help Attack.clear().
Changed ranging of monsters in the special classid list (like summoners) to range from current location instead of original location. This will help situations where you run after something and run into a summoner.
Changed the sortMonster function to return 0 if two monsters are both in the special id list and normal (not a boss). Hopefully this will help (along with the stability of SpiderMonkey's sort) to keep you attacking the same monster when appropriate.
Prioritizes units in the classid list that are bosses higher so that uniques like Bishibosh will be killed first.