Pather.useWaypoint: ensure a waypoint menu is opened if "check" is "true"
When two or more bots are trying to get a waypoint simultaneously and "check" is "true", some may fail to get menu opened, hence, interact directly. D2GS derivatives treat such a behavior as a cheat and block the action. Also, account may get a permanent ban, for an instance, Rubattle.net is the case. At Rubattle.net, "check" in the "true" state is a must. So if you wanna avoid unexpected bans at such nasty servers, just change line 796 to "if (true) {".

D2GS output: 
2014/11/1 23:40:32.140 [cheat] minx-nec(*minx-nec)@Hcs9: packet data dump: 49120000006b000000