Distance check
Do not need to check the distance twice, also having the check there bound the check to a single passed unit parameter. The desired effect is to have it continually use CE as long as it can damage monster packs, the original logic had it use CE twice for a single monster unit. Then wait for the next loop to use it again.