Update Necromancer.js
With the addition of using CorpseExplosion during normal attacks I re-aligned the changes in the original explodecorpses function to take advantage of that. i.e. Stop trying to do extra work. Will now just use up the corpses until either the monster dies or we run out of bodies in which case the normal attack sequence will resume.