[TeamBaal] First commit of Team Baal. Added a script that can throw a teamed baalparty. However also solo it have mature improvements
It speeds up a run, it communicates between all the profiles running on the same machine, and have minimum configuration
It removed 30 seconds a run from switching my team between Baal and BaalAssistant/helper.
However, there are more improvements to be made, and this is the first setup. More to come