Add BattleOrders tardy check and quit/continue vars
This removes the unnecessary delay when someone rejoins a game or joins
late and the time for battle orders has passed. The bot will now check
for other player locations and just continue on instead of waiting the
BattleOrder timeout. Instead of quitting if the timeout is reached,
now the character can optionally just continue on with the run. These
two changes combined make for smoother battle order control.