Add SealOrder handling to Diablo.js
SealOrder is checked to see if a specific order has been defined. The
default order is the old order, viz/seis/infector. If a user defined
order is in place, that order is followed. The order is respected even
if less than three seals are listed. If Diablo is not found, all the
seals are checked in the default order.