Add NoTele param [default: false]
With this param set to true, a character will not teleport during
pathing, even if the teleport skill is available.

This is helpful for chars with low mana management (low max or regen).
Without this, if a char runs out of mana too quickly during pathing,
they'll either wait for regen (halt, and wait) or will visit town to
refill at npc way too often.